requirements.txt
# --- Core agent orchestration ---
langgraph
langchain-core
langchain-openai
litellm

# --- MCP client ---
mcp

# --- A2A ---
a2a-sdk[telemetry]

# --- FastAPI / Starlette / HTTP & streaming ---
fastapi
starlette
uvicorn[standard]
orjson
python-multipart
pydantic
pydantic-settings
PyYAML
typer
httpx
sse-starlette
aiofiles
jinja2

# --- AuthN/AuthZ & security ---
Authlib
python-jose[cryptography]
passlib[bcrypt]
casbin
python-dotenv

# --- Observability (OTel + Prometheus) ---
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-grpc
prometheus-client
prometheus-fastapi-instrumentator
opentelemetry-instrumentation-asgi


# --- Storage / persistence ---
psycopg[binary]
redis

# --- Optional transport ---
grpcio
grpcio-tools

# --- QoL / utilities ---
tenacity
structlog
typing-extensions
